форма авторизации вынесена в отдельный компонент, написаны тесты для него
