+++ project/src/pages/login-screen/login-screen.test.tsx
@@ -11,23 +11,8 @@ import AuthorisationStatus from '../../types/authorisation-status';
 import { Route, Routes } from 'react-router-dom';
 import { changeCity } from '../../store/application-process/application-process';
 import CitiesName from '../../types/cities-name';
-import { createAPI } from '../../services/api';
-import MockAdapter from 'axios-mock-adapter';
-import thunk from 'redux-thunk';
-import AdditionalURL from '../../types/additional-url';
-import { State } from '../../types/state';
-import { Action } from 'redux';
-import { logIn } from '../../store/thunk-actions';
-import { generateUser } from '../../utils/mocks';
 
-const MOCK_EMAIL = 'someEmail@mockEmail.com';
-const MOCK_PASSWORD = 'password123';
-const SIGN_IN_ELEMENTS_COUNT = 2;
-
-const api = createAPI();
-const mockAPI = new MockAdapter(api);
-const middlewares = [thunk.withExtraArgument(api)];
-const mockStore = configureMockStore<State, Action<string>>(middlewares);
+const mockStore = configureMockStore();
 
 const mockStoreWithAuthorisation = (flag = true) => mockStore({
   [NameSpace.User]:{
@@ -53,25 +38,20 @@ describe('Component: LoginScreen', ()=>{
 
   beforeEach(() => history.push(AppRoute.Login));
 
-  it('should render when an unauthorised user navigate to "login" url', async()=>{
+  it('should render when an unauthorised user navigate to "login" url', ()=>{
     const store = mockStoreWithAuthorisation(false);
-
     render(fakeLoginScreen(store));
 
-    const signInElements = screen.getAllByText(/Sign in/i);
-    expect(signInElements.length).toBe(SIGN_IN_ELEMENTS_COUNT);
-    signInElements.forEach((element)=>expect(element).toBeInTheDocument());
+    const headerElement = screen.getByRole('heading');
+    const buttonElemnent = screen.getByRole('button');
 
-    expect(screen.getByTestId('changing-city-link')).toBeInTheDocument();
+    expect(headerElement).toHaveTextContent(/Sign in/i);
+    expect(headerElement).toBeInTheDocument();
     expect(screen.getByLabelText(/E-mail/i)).toBeInTheDocument();
     expect(screen.getByLabelText(/Password/i)).toBeInTheDocument();
-    expect(screen.getByRole('button')).toBeInTheDocument();
-
-    await userEvent.type(screen.getByTestId('email'), MOCK_EMAIL);
-    await userEvent.type(screen.getByTestId('password'), MOCK_PASSWORD);
-
-    expect(screen.getByDisplayValue(MOCK_EMAIL)).toBeInTheDocument();
-    expect(screen.getByDisplayValue(MOCK_PASSWORD)).toBeInTheDocument();
+    expect(buttonElemnent).toHaveTextContent(/Sign in/i);
+    expect(buttonElemnent).toBeInTheDocument();
+    expect(screen.getByTestId('changing-city-link')).toBeInTheDocument();
   });
 
   it('should redirect to "/" when user clicks the "changing-city-link"', async()=>{
@@ -96,34 +76,6 @@ describe('Component: LoginScreen', ()=>{
     expect(store.getActions()).toEqual([
       changeCity(chosenCity)
     ]);
-
-  });
-
-  it('should dispatch the "logIn" action when user is getting authorised', async()=>{
-    const user = {
-      ...generateUser(),
-      email: MOCK_EMAIL
-    };
-    mockAPI
-      .onPost(AdditionalURL.Login)
-      .reply(200, user);
-
-    const store = mockStoreWithAuthorisation(false);
-    Storage.prototype.setItem = jest.fn();
-
-    render(fakeLoginScreen(store));
-
-    await userEvent.type(screen.getByTestId('email'), MOCK_EMAIL);
-    await userEvent.type(screen.getByTestId('password'), MOCK_PASSWORD);
-
-    expect(store.getActions()).toEqual([]);
-
-    await userEvent.click(screen.getByRole('button'));
-
-    const actions = store.getActions().map(({type}) => type);
-
-    expect(actions.includes(logIn.pending.type)).toBe(true);
-
   });
 
   it('should not render, but redirect to "/" when user is authorised', ()=>{
